Refactor SQL dataframe operation
Piggybacking on the recent fixes to the sql decorator, we can now
simplify the SQL dataframe decorator to only inherit from
DecoratedOperator and take advantage of functions created in the
Database class.

(cherry picked from commit b17bca4378ee7e406865d358e22b3c796db2918c)
